--- Project/AcceptanceTests/Bridge/SystemBridge.java
@@ -1,4 +1,0 @@
-package Bridge;
-
-public interface SystemBridge {
-}
