--- Project/AcceptanceTests/DataObjects/Guest.java
@@ -1,12 +1,0 @@
-package DataObjects;
-
-public class Guest extends User {
-
-    public Guest(int ID){
-        this.ID = ID;
-    }
-
-    public int getID() {
-        return ID;
-    }
-}
