--- Project/AcceptanceTests/DataObjects/Product.java
@@ -1,12 +1,0 @@
-package DataObjects;
-
-public class Product {
-    public String name;
-    public String manufacturer;
-
-    public Product(String name, String manufacturer){
-        this.name = name;
-        this.manufacturer = manufacturer;
-    }
-
-}
