--- Project/AcceptanceTests/DataObjects/ProductFilter.java
@@ -1,5 +1,0 @@
-package DataObjects;
-
-public interface ProductFilter {
-    boolean filter(ProductInShop product);
-}
