--- Project/AcceptanceTests/DataObjects/RegistrationInfo.java
@@ -1,11 +1,0 @@
-package DataObjects;
-
-public class RegistrationInfo {
-    public String username;
-    public String password;
-
-    public RegistrationInfo(String username,String password){
-        this.username = username;
-        this.password = password;
-    }
-}
