--- Project/AcceptanceTests/DataObjects/ShopFilter.java
@@ -1,5 +1,0 @@
-package DataObjects;
-
-public interface ShopFilter {
-    boolean filter(Shop shop);
-}
