--- Project/AcceptanceTests/DataObjects/User.java
@@ -1,12 +1,0 @@
-package DataObjects;
-
-import java.util.List;
-
-public abstract class User {
-    public int ID;
-    public List<String> notifications;
-    public int numOfNotifications() {
-        return notifications.size();
-    }
-
-}
