--- Project/AcceptanceTests/Tests/GuestTests.java
@@ -1,9 +1,0 @@
-package Tests;
-
-import DataObjects.User;
-
-public class GuestTests extends UserTests {
-    public User enter(){
-        return UserTests.userBridge.visit();
-    }
-}
