+++ Project/Tests/AcceptanceTests/Bridge/ShopsBridge.java
@@ -1,0 +1,16 @@
+package AcceptanceTests.Bridge;
+
+import AcceptanceTests.DataObjects.Product;
+import AcceptanceTests.DataObjects.Shop;
+import AcceptanceTests.DataObjects.User;
+
+
+public interface ShopsBridge {
+    boolean addShop(Shop shop, User founder);
+
+    void addProductToShop(int shopID, Product product,int productID, double rating, int quantity, double price);
+
+    void addManager(int shopID, int userID);
+
+    void appointOwner(int shopID, int appointerID, int appointeeID);
+}
