+++ Project/Tests/AcceptanceTests/Bridge/SystemBridge.java
@@ -1,0 +1,4 @@
+package AcceptanceTests.Bridge;
+
+public interface SystemBridge {
+}
