+++ Project/Tests/AcceptanceTests/DataObjects/Guest.java
@@ -1,0 +1,12 @@
+package AcceptanceTests.DataObjects;
+
+public class Guest extends User {
+
+    public Guest(int ID){
+        this.ID = ID;
+    }
+
+    public int getID() {
+        return ID;
+    }
+}
