+++ Project/Tests/AcceptanceTests/DataObjects/Product.java
@@ -1,0 +1,12 @@
+package AcceptanceTests.DataObjects;
+
+public class Product {
+    public String name;
+    public String manufacturer;
+
+    public Product(String name, String manufacturer){
+        this.name = name;
+        this.manufacturer = manufacturer;
+    }
+
+}
