+++ Project/Tests/AcceptanceTests/DataObjects/ProductFilter.java
@@ -1,0 +1,5 @@
+package AcceptanceTests.DataObjects;
+
+public interface ProductFilter {
+    boolean filter(ProductInShop product);
+}
