+++ Project/Tests/AcceptanceTests/DataObjects/RegistrationInfo.java
@@ -1,0 +1,11 @@
+package AcceptanceTests.DataObjects;
+
+public class RegistrationInfo {
+    public String username;
+    public String password;
+
+    public RegistrationInfo(String username,String password){
+        this.username = username;
+        this.password = password;
+    }
+}
