+++ Project/Tests/AcceptanceTests/DataObjects/ShopFilter.java
@@ -1,0 +1,5 @@
+package AcceptanceTests.DataObjects;
+
+public interface ShopFilter {
+    boolean filter(Shop shop);
+}
