+++ Project/Tests/AcceptanceTests/DataObjects/User.java
@@ -1,0 +1,12 @@
+package AcceptanceTests.DataObjects;
+
+import java.util.List;
+
+public abstract class User {
+    public int ID;
+    public List<String> notifications;
+    public int numOfNotifications() {
+        return notifications.size();
+    }
+
+}
