+++ Project/Tests/AcceptanceTests/Tests/GuestTests.java
@@ -1,0 +1,9 @@
+package AcceptanceTests.Tests;
+
+import AcceptanceTests.DataObjects.User;
+
+public class GuestTests extends UserTests {
+    public User enter(){
+        return userBridge.visit();
+    }
+}
